:recycle: refactor some controller methods to resource params
